transaction_consumer_enriched/pyproject.toml
[project]
name = "transaction-consumer-enriched"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka==2.10.1",
    "minio==7.2.15",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.0",
    "sqlalchemy==2.0.36",
    "pydantic==2.6.2",
    "python-dateutil",
]
